+++ flutter_inappwebview_android/lib/src/main.dart
@@ -16,3 +16,5 @@ export 'proxy_controller.dart' hide InternalProxyController;
 export 'webview_asset_loader.dart';
 export 'tracing_controller.dart' hide InternalTracingController;
 export 'process_global_config.dart' hide InternalProcessGlobalConfig;
+
+export 'webview_profile.dart';